+++ jvm/src/test/scala/eval/EvaluationSpec.scala
@@ -179,7 +179,7 @@ class EvaluationSpec extends FunSpec with Matchers with ExpressionTreeCodec {
       case Valid(t) => t.bands.head.get(0, 0) should be (1)
       case i@Invalid(_) => fail(s"$i")
     }
-    interpreter(FocalSlope(List(IntArrayTile(1 to 100 toArray, 10, 10)), TargetCell.All)).as[MultibandTile] match {
+    interpreter(FocalSlope(List(IntArrayTile(1 to 100 toArray, 10, 10)), None, TargetCell.All)).as[MultibandTile] match {
       case Valid(t) => t.bands.head.get(5, 5) should be (10)
       case i@Invalid(_) => fail(s"$i")
     }
@@ -205,7 +205,7 @@ class EvaluationSpec extends FunSpec with Matchers with ExpressionTreeCodec {
     val hillshadeProjectedRaster =
       ProjectedRaster(Raster(MultibandTile(hillshadeTile), hillshadeE), WebMercator)
 
-    interpreter(FocalHillshade(List(RasterLit(hillshadeProjectedRaster)), 315, 45, TargetCell.All)).as[MultibandTile] match {
+    interpreter(FocalHillshade(List(RasterLit(hillshadeProjectedRaster)), 315, 45, None, TargetCell.All)).as[MultibandTile] match {
       case Valid(t) => t.bands.head.get(2, 2) should be (77)
       case i@Invalid(_) => fail(s"$i")
     }